+++ game-engine/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod renderer;
+pub mod physics;
 pub mod context;
 pub use context::Context;
