+++ game-engine/src/physics/box/mesh.rs
@@ -1,0 +1,115 @@
+use crate::{physics::{Vec3, Ray, Tri, Quaternion, PhysicsMaterial}, renderer::Transform};
+
+use super::BoxColider;
+
+// TODO FIX using get_verts
+pub fn get_vertex(w: &Vec3, t: &Transform, c: &BoxColider) -> Vec<Vec3> {
+    let s = c.scale * t.scale;
+    let r = t.rotation * c.local_rotation;
+    let mut vec: Vec<Vec3> = Vec::with_capacity(8);
+
+    for x in [-1.0, 1.0] {
+        for y in [-1.0, 1.0] {
+            for z in [-1.0, 1.0] {
+                vec.push(w + r * Vec3::new(s.x * x, s.y * y, s.z * z))
+            }
+        }
+    }
+
+    vec
+}
+
+#[must_use]
+/// in binary order, aka v000 v001 v010, not rotated where v000 is min and v111 is max
+pub fn get_verts(t: &Transform, c: &BoxColider) -> [Vec3; 8] {
+    let c = t.scale * c.scale;
+    let v111 = c;
+    let v000 = -c;
+
+    let v001 = Vec3::new(v000.x, v000.y, v111.z);
+    let v010 = Vec3::new(v000.x, v111.y, v000.z);
+    let v100 = Vec3::new(v111.x, v000.y, v000.z);
+
+    let v011 = Vec3::new(v000.x, v111.y, v111.z);
+    let v110 = Vec3::new(v111.x, v111.y, v000.z);
+    let v101 = Vec3::new(v111.x, v000.y, v111.z);
+
+    [v000, v001, v010, v011, v100, v101, v110, v111]
+}
+
+#[test]
+fn test_get_verts() {
+    let scale = Vec3::new(2.0, 1.0, 10.0);
+
+    let t = Transform {
+        position: Vec3::zero(),
+        rotation: Quaternion::identity(),
+        scale,
+    };
+
+    let material = PhysicsMaterial {
+        friction: 1.0,
+        restfullness: 1.0,
+    };
+
+    let c = BoxColider::new(Vec3::one(), material);
+    let verts = get_verts(&t, &c);
+
+    for x in [-1, 1] {
+        for y in [-1, 1] {
+            for z in [-1, 1] {
+                assert!(verts.contains(&Vec3::new(
+                    scale.x * x as f32,
+                    scale.y * y as f32,
+                    scale.z * z as f32,
+                )))
+            }
+        }
+    }
+}
+
+#[must_use]
+pub fn get_rays_for_box(verts: &[Vec3; 8]) -> [Ray; 12] {
+    let [v000, v001, v010, v011, v100, v101, v110, _v111] = *verts;
+    [
+        Ray::new(v000, Vec3::unit_x()),
+        Ray::new(v001, Vec3::unit_x()),
+        Ray::new(v010, Vec3::unit_x()),
+        Ray::new(v011, Vec3::unit_x()),
+        Ray::new(v000, Vec3::unit_y()),
+        Ray::new(v001, Vec3::unit_y()),
+        Ray::new(v100, Vec3::unit_y()),
+        Ray::new(v101, Vec3::unit_y()),
+        Ray::new(v000, Vec3::unit_z()),
+        Ray::new(v010, Vec3::unit_z()),
+        Ray::new(v100, Vec3::unit_z()),
+        Ray::new(v110, Vec3::unit_z()),
+    ]
+}
+
+#[must_use]
+pub fn get_tris_for_box(verts: &[Vec3; 8]) -> [Tri; 12] {
+    let [v000, v001, v010, v011, v100, v101, v110, v111] = *verts;
+
+    // counter clockwise
+    [
+        // x face
+        [v101, v100, v110],
+        [v101, v110, v111],
+        // -x face
+        [v000, v001, v011],
+        [v000, v011, v010],
+        // y face
+        [v111, v010, v011],
+        [v010, v111, v110],
+        // -y face
+        [v000, v101, v001],
+        [v000, v100, v101],
+        // z face
+        [v001, v101, v111],
+        [v001, v111, v011],
+        //-z face
+        [v000, v010, v100],
+        [v010, v110, v100],
+    ]
+}
